+++ Core/Executor/ContentManager.php
@@ -317,7 +317,7 @@ class ContentManager extends RepositoryExecutor implements MigrationGeneratorInt
     protected function matchContents($action)
     {
         if (!isset($this->dsl['object_id']) && !isset($this->dsl['remote_id']) && !isset($this->dsl['match'])) {
-            throw new \Exception("The ID or remote ID of an object or a Match Condition is required to $action a new location.");
+            throw new \Exception("The id or remote id of an object or a match condition is required to $action a location");
         }
 
         // Backwards compat
@@ -329,18 +329,8 @@ class ContentManager extends RepositoryExecutor implements MigrationGeneratorInt
             }
         }
 
-        $match = $this->dsl['match'];
-
         // convert the references passed in the match
-        foreach ($match as $condition => $values) {
-            if (is_array($values)) {
-                foreach ($values as $position => $value) {
-                    $match[$condition][$position] = $this->referenceResolver->resolveReference($value);
-                }
-            } else {
-                $match[$condition] = $this->referenceResolver->resolveReference($values);
-            }
-        }
+        $match = $this->resolveReferencesRecursively($this->dsl['match']);
 
         return $this->contentMatcher->match($match);
     }